Type Alias linera_alloy_transport::TransportError

pub type TransportError<ErrResp = Box<RawValue>> = RpcError<TransportErrorKind, ErrResp>;

A transport error is an RpcError containing a [TransportErrorKind].

Aliased Type

enum TransportError<ErrResp = Box<RawValue>> {
    ErrorResp(ErrorPayload<ErrResp>),
    NullResp,
    UnsupportedFeature(&'static str),
    LocalUsageError(Box<dyn Error + Send + Sync>),
    SerError(Error),
    DeserError {
        err: Error,
        text: String,
    },
    Transport(TransportErrorKind),
}

Variants

ErrorResp(ErrorPayload<ErrResp>)

Server returned an error response.

NullResp

Server returned a null response when a non-null response was expected.

UnsupportedFeature(&'static str)

Rpc server returned an unsupported feature.

LocalUsageError(Box<dyn Error + Send + Sync>)

Returned when a local pre-processing step fails. This allows custom errors from local signers or request pre-processors.

SerError(Error)

JSON serialization error.

Tuple Fields

0: Error

The underlying serde_json error.

DeserError

JSON deserialization error.

Fields

err: Error

The underlying serde_json error.

text: String

For deser errors, the text that failed to deserialize.

Transport(TransportErrorKind)

Transport error.

This variant is used when the error occurs during communication.

Tuple Fields

0: TransportErrorKind

The underlying transport error.